Added flambe.script.
The script/action system is analogous to Flashbang tasks. A Script
component is added to an Entity, and Actions are added to the Script.

The word "script" here is in the context of theatre, which game engine
authors are crazy about, and fits nicely with all the Scenes, Stages and
Directors.