Disposer manages Disposables.
This fulfills the same role as EventHandlerManager (from Aspirin), but a
lot easier to integrate with entities and can manage any Disposable, not
just SignalConnections.

Not sure about having this live in the top level 'flambe' package. I
think we'll end up having a package containing a toolbox of handy
components later.