Add System.uncaughtError signal.
This doesn't include a stack trace or line number, but it suffices as a first line of defence
against bugs that only happen in the wild. In my game, I've got this hooked up to Google Analytics,
which will also send an email alert if activity passes a certain threshold.