Remove Transform and merge its functionality into Sprite.
I originally thought Transform might be a useful standalone component even
without a sprite, but that has never been the case. It's not worth the extra
indirection when you just want to figure out how to move a sprite around.

So, Sprite grows a tiny bit, but I'm happy with this. I'm also ok with Sprite
having a bunch of subclasses, even though this is a component based engine. I'm
finding components to be highly useful, but a "pure" component system is way too
much freaking indirection to be able to get anything done quickly. Also,
components seem to have much more benefit for game logic rather than game
engine architecture.