Replace Property with Value and AnimatedFloat.
I wanted a reactive type for values that aren't necessarily animated over time,
and make sense for non-numbers, which is where Value comes in.

And since in practice the main thing that gets animated over time are Floats,
the parametric type on Property and Behavior was kinda cumbersome/broken.
flambe.animation is restricted to just Floats for now.