Add HTML mouse listeners to window.
Previously dragging and releasing outside of the stage would make
Pointer.isDown() return bad values. This makes things work as they do in Flash,
where dragging outside of the stage will keep emitting Pointer.move until the
mouse is released.

Actually, this seems a little nicer than in Flash, as Pointer.move will always
be emitted even while outside of the stage. So mouse followers will always work
as expected, instead of stopping at the edge of the stage as they do in Flash.

Touch events shouldn't require a change, touchmove is already emitted when
dragging outside of the canvas.