Use a Key enum instead of native key codes.
Not only is this easier to use, but it insulates users from platforms
that have different key codes. Luckily Flash and JS seem fairly
consistent, so there's no need for that yet.

I've only included the keys that Flash and JS actually emit. This could
probably be expanded a bit.