Allow both landscape orientations.
This is new to AIR 3.3, but it's still quirky. If you lockOrientation to
landscape while in the rotated-right landscape, AIR will flip to the
rotated-left landscape. Maybe they'll get it right in a later version?

This could be solved by manually listening to ORIENTATION_CHANGING and
selectively cancelling, but I've had enough of messing with this for
now.