Some fixes for pointer listener leakage.
This is still a problem. The difference between a 'removed' and
'disposed' component is confusing and error prone, and something I'm
hoping to simplify soon.

The whole interactive sprite thing also needs an overhaul. It should be
possible to add a listener to a parent sprite to be notified when any of
its children are clicked, but that's not really the case at the moment.