Add an inspector to help optimize and debug rendering.
This thing is crazy useful! The inspector will log all draw calls, and
spit out a screenshot of the frame at each step. This is awesome for
optimizing texture swapping and overdraw, and helps to visualize how
something is being rendered. Press CTRL-I to use it, in an HTML debug
build.