Handling of properties possibly not supported on all platforms.
Introduced possible concept of a holder class for a single float value to help avoid accidental usage when a particular property is not available on a specific platform. Class is only instanced when the property is available. NOTE: This version needs extensive testing, and parts of it are preliminary, as some of it may change based on how AIR/NME are able to conform to this API.

Signed-off-by: Bradley Baysinger <bradley@seven2.com>