Accelerometer API improvements via @braddelcat, with some adjustments.
Mainly name fiddling, and accelerationIncludesGravity is a flag on the
emitted acceleration vector instead of a separate signal. Which is a bit
nicer to deal with when writing games that need to work in both cases.

Also renamed Orientation to Attitude as suggested to avoid ambiguity
with flambe.display.Orientation.