Disable mipmapping.
Keeping mipmaps up to date is problematic with render-to-texture. If the
base level of a texture is updated, we have no way of knowing that its
mipmaps should be regenerated.

Later on, it should be possible to mark a texture as invalid when
rendered to, and lazily regenerate the mipmaps as needed. Also, a public
API to request mipmapping on a Texture.