Extract flambe.server out into its own lib.
It lives at https://github.com/aduros/flambe-server now.

I suspect Everybody Runs is the only project that uses it. It's not
really maintained at the moment, and was a bit out of place with the
rest of Flambe core.

I have a bunch of application-server type stuff written for ER that I
might clean up and move over to that project someday.