Detect pointer properly when canvas is scaled by the page
If someone were trying to manually scale the canvas via CSS (i.e. to
fill the window with a flambe_disable_autoresize game), the pointer
coordinates would fail to be scaled to match. This should allow for
arbitrary custom resizing to be applied, and still work with default
situations (including retina compensation).