Benchmark test
Benchmark rates the performance of the device based on how long it
takes the app to add 5000 empty entities to the stage.  The higher the
number the better the performance.

Call Benchmark.start(functionName). functionName will be called when
the test is complete.  You can also call Benchmark.results() to get the
value

From my own tests
- low performance: under 2000
- high performance: above 2000