Tiny tweak.
I like to explicitly initialize member vars, because Haxe JS doesn't do
it automatically. This doesn't (usually) result in bugs, but there are
performance issues with adding new fields to JS objects outside of the
constructor.

It would be great if Haxe did this automatically. Maybe I'll look into
that.