Add dox templates.
Hell is frozen over! Closes #259.

There's still a couple little things needed to get it playing nicely
with Dox's new theme system. Mainly a fix to dpeek/dox#44. The "view
source" links also need to be cleaned up, because it's hacked together
with hardcoded paths right now.

That said, it should be in pretty good shape for other projects to fork.